pure r cmd check compiler bliss

+++ src/doc.h
@@ -35,7 +35,7 @@ public:
   void gen_biterms(vector<Biterm>& bs, int win = 15) const {
 	if (ws.size() < 2) return;
 	
-	for (int i = 0; i < ws.size()-1; ++i) 
+	for (int i = 0; i < (int)ws.size()-1; ++i) 
 	  for (int j = i+1; j < min(i + win, int(ws.size())); ++j) 
 		bs.push_back( Biterm(ws[i], ws[j]) );
   }
